Nodes may be "strongly held."
Nodes have a .stronglyHeld attribute.  A node is strongly held if any of the
nodes it .sendsTo is strongly held.  EventStream.pinE() creates a pass-through
node that is always strongly held.

When extractEventStaticE receives an event and is not strongly held, it
removes its event listener and does not forward the event.